Separeted Graph inform from event logging.
We used to have an older ALEventLogger that showed events in the console
log. The new graph was added to the same component. But as a result, the
main content of the app was pushed way down.

In this step, separeted the graph from event logger (will keep that for
debugging if every needed).
