+++ packages/hyperion-react-testapp/src/App.tsx
@@ -2,21 +2,18 @@
  * Copyright (c) Meta Platforms, Inc. and affiliates. All Rights Reserved.
  */
 
+import { LocalStoragePersistentData } from '@hyperion/hyperion-util/src/PersistentData';
 import React, { ChangeEventHandler, useCallback, useState } from 'react';
 import './App.css';
-import LargeComp from './component/LargeComponent';
-import Counter from "./component/Counter";
-import NestedComponent from './component/NestedComponent';
-import { PortalBodyContainerComponent } from './component/PortalComponent';
+import ALGraphView from './component/ALGraphView';
 import DynamicSvgComponent from './component/DynamicSvgComponent';
 import ElementNameComponent from './component/ElementNameComponent';
-import TextComponent from './component/TextComponent';
-import RecursiveRuncComponent from "./component/RecursiveFuncComponent";
-import { ElementTextTooltip } from "@hyperion/hyperion-autologging-visualizer/src/component/ElementTextTooltip.react";
-import { SyncChannel } from './Channel';
+import LargeComp from './component/LargeComponent';
+import NestedComponent from './component/NestedComponent';
 import NonInteractiveSurfaceComponent from './component/NonInteractiveSurfaceComponent';
-import ALEventLogger from './component/ALEventLogger';
-import { LocalStoragePersistentData } from '@hyperion/hyperion-util/src/PersistentData';
+import { PortalBodyContainerComponent } from './component/PortalComponent';
+import RecursiveRuncComponent from "./component/RecursiveFuncComponent";
+import TextComponent from './component/TextComponent';
 
 function InitComp() {
   const [count, setCount] = React.useState(0);
@@ -80,7 +77,7 @@ function App() {
 
   return (
     <div className="App">
-      <ALEventLogger />
+      <ALGraphView />
       <label htmlFor='testSelector'>Select a mode:</label>
       <select onChange={onChange} value={mode} id='testSelector' aria-label='Mode Selector'>
         {Object.keys(Modes).map(key => <option key={key} value={key}>{key}</option>)}
